pyproject.toml
[project]
name = "backend"
version = "0.0.1"
description = "Default template for PDM package"
authors = [
    {name = "", email = "matheus@mlopes.xyz"},
]
dependencies = ["django>=5.1.5", "djangorestframework>=3.15.2"]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
